+++ PartsManagementSystem/Model/Protocols/Vendor.swift
@@ -6,10 +6,12 @@
 //  Copyright © 2018 Łukasz Sypniewski. All rights reserved.
 //
 
+import Foundation
+
 protocol Vendor: AnyObject {
     var name: String { get }
     var address: AddressProtocol { get set }
     var availableProducts: [ManageableProduct] { get set}
-    var commission: Double { get }
+    var commission: Decimal { get }
     static var priceCalculator: PriceCalculation { get set }
 }
