+++ PartsManagementSystem/Model/Shop.swift
@@ -6,14 +6,16 @@
 //  Copyright © 2018 Łukasz Sypniewski. All rights reserved.
 //
 
+import Foundation
+
 class Shop: Vendor {
     var name: String
     var address: AddressProtocol
     var availableProducts: [ManageableProduct] = [ManageableProduct]()
-    var commission: Double
+    var commission: Decimal
     static var priceCalculator: PriceCalculation { set { Shop.priceCalculator = newValue } get { return Shop.priceCalculator } }
     
-    init(name: String, address: Address, commission: Double) {
+    init(name: String, address: Address, commission: Decimal) {
         self.name = name
         self.address = address
         self.commission = commission
